src/main/java/cn/yfbai/shopbackend/service/ProductService.java
package cn.yfbai.shopbackend.service;

import cn.yfbai.shopbackend.entity.Product;
import cn.yfbai.shopbackend.repository.ProductRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.math.BigDecimal;
import java.util.Arrays;
import java.util.List;

@Service
public class ProductService {

    @Autowired
    private ProductRepository productRepository;

    public List<Product> getProducts() {
        return productRepository.findAll();
    }
}
